--- a/reactor-aeron-benchmarks/src/main/java/reactor/aeron/rsocket/aeron/RSocketAeronServerTps.java
+++ b/reactor-aeron-benchmarks/src/main/java/reactor/aeron/rsocket/aeron/RSocketAeronServerTps.java
@@ -10,13 +10,11 @@ import io.rsocket.RSocketFactory;
 import io.rsocket.reactor.aeron.AeronServerTransport;
 import io.rsocket.util.ByteBufPayload;
 import java.util.Random;
-import java.util.concurrent.Callable;
 import reactor.aeron.AeronResources;
 import reactor.aeron.AeronServer;
 import reactor.aeron.Configurations;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Schedulers;
 
 public final class RSocketAeronServerTps {
 
@@ -58,12 +56,11 @@ public final class RSocketAeronServerTps {
                       public Flux<Payload> requestStream(Payload payload) {
                         payload.release();
 
-                        Callable<Payload> payloadCallable =
-                            () -> ByteBufPayload.create(BUFFER.retainedSlice());
+                        long msgNum = Configurations.NUMBER_OF_MESSAGES;
+                        System.out.println("streaming " + msgNum + " messages ...");
 
-                        return Mono.fromCallable(payloadCallable)
-                            .subscribeOn(Schedulers.parallel())
-                            .repeat(Configurations.NUMBER_OF_MESSAGES);
+                        return Flux.range(0, Integer.MAX_VALUE)
+                            .map(i -> ByteBufPayload.create(BUFFER.retainedSlice()));
                       }
                     }))
         .transport(