Some intermediate benchmark results:

before updating to the new rsocket version (0.11.18)

`rsocket-netty-client-tps.sh`
```
176495.9 msgs/sec, 172.3593 MB/sec, totals 494882 messages 483 MB payloads
178748.8 msgs/sec, 174.5594 MB/sec, totals 673631 messages 657 MB payloads
174823.9 msgs/sec, 170.7265 MB/sec, totals 848456 messages 828 MB payloads
171653.9 msgs/sec, 167.6307 MB/sec, totals 1020109 messages 996 MB payloads
175918.2 msgs/sec, 171.7951 MB/sec, totals 1196027 messages 1167 MB payloads
172902.0 msgs/sec, 168.8496 MB/sec, totals 1368929 messages 1336 MB payloads
174138.5 msgs/sec, 170.0571 MB/sec, totals 1543067 messages 1506 MB payloads
175275.3 msgs/sec, 171.1673 MB/sec, totals 1718343 messages 1678 MB payloads
```

`rsocket-aeron-client-tps.sh`
```
140178.8 msgs/sec, 273.7867 MB/sec, totals 488312 messages 953 MB payloads
143011.8 msgs/sec, 279.3199 MB/sec, totals 631323 messages 1233 MB payloads
138661.9 msgs/sec, 270.8240 MB/sec, totals 769985 messages 1503 MB payloads
149466.4 msgs/sec, 291.9265 MB/sec, totals 919451 messages 1795 MB payloads
126756.2 msgs/sec, 247.5707 MB/sec, totals 1046207 messages 2043 MB payloads
148149.5 msgs/sec, 289.3544 MB/sec, totals 1194357 messages 2332 MB payloads
150223.5 msgs/sec, 293.4053 MB/sec, totals 1344581 messages 2626 MB payloads
```

`rsocket-aeron-client-simple-tps.sh`
```
143889.5 msgs/sec, 140.5171 MB/sec, totals 531648 messages 519 MB payloads
161002.6 msgs/sec, 157.2291 MB/sec, totals 692650 messages 676 MB payloads
157325.9 msgs/sec, 153.6386 MB/sec, totals 849976 messages 830 MB payloads
162753.9 msgs/sec, 158.9393 MB/sec, totals 1012731 messages 988 MB payloads
157834.3 msgs/sec, 154.1350 MB/sec, totals 1170565 messages 1143 MB payloads
160819.9 msgs/sec, 157.0507 MB/sec, totals 1331384 messages 1300 MB payloads
151218.8 msgs/sec, 147.6746 MB/sec, totals 1482603 messages 1447 MB payloads
135899.8 msgs/sec, 132.7146 MB/sec, totals 1618503 messages 1580 MB payloads
160565.5 msgs/sec, 156.8022 MB/sec, totals 1779068 messages 1737 MB payloads 
```


after:


`rsocket-netty-client-tps.sh`
```
930500.4 msgs/sec, 908.6908 MB/sec, totals 5423750 messages 5296 MB payloads
975143.8 msgs/sec, 952.2898 MB/sec, totals 6398857 messages 6248 MB payloads
938002.1 msgs/sec, 916.0167 MB/sec, totals 7336860 messages 7164 MB payloads
961824.6 msgs/sec, 939.2818 MB/sec, totals 8298685 messages 8104 MB payloads
935558.4 msgs/sec, 913.6313 MB/sec, totals 9234246 messages 9017 MB payloads
908960.8 msgs/sec, 887.6571 MB/sec, totals 10143209 messages 9905 MB payloads
922026.4 msgs/sec, 900.4164 MB/sec, totals 11065230 messages 10805 MB payloads
941152.9 msgs/sec, 919.0946 MB/sec, totals 12006419 messages 11725 MB payloads
922202.4 msgs/sec, 900.5883 MB/sec, totals 12928590 messages 12625 MB payloads
915365.3 msgs/sec, 893.9124 MB/sec, totals 13843957 messages 13519 MB payloads
952970.5 msgs/sec, 930.6342 MB/sec, totals 14796965 messages 14450 MB payloads
920271.1 msgs/sec, 898.7032 MB/sec, totals 15717193 messages 15348 MB payloads
963755.9 msgs/sec, 941.1669 MB/sec, totals 16680960 messages 16290 MB payloads
896888.2 msgs/sec, 875.8684 MB/sec, totals 17577842 messages 17165 MB payloads
818848.8 msgs/sec, 799.6561 MB/sec, totals 18396689 messages 17965 MB payloads
922308.5 msgs/sec, 900.6929 MB/sec, totals 19319053 messages 18866 MB payloads
924455.9 msgs/sec, 902.7880 MB/sec, totals 20243459 messages 19769 MB payloads
929645.1 msgs/sec, 907.8565 MB/sec, totals 21173103 messages 20676 MB payloads
924683.7 msgs/sec, 903.0114 MB/sec, totals 22097789 messages 21579 MB payloads
941980.1 msgs/sec, 919.9025 MB/sec, totals 23039759 messages 22499 MB payloads
873064.6 msgs/sec, 852.6031 MB/sec, totals 23912863 messages 23352 MB payloads
916554.3 msgs/sec, 895.0716 MB/sec, totals 24829379 messages 24247 MB payloads
935223.3 msgs/sec, 913.3050 MB/sec, totals 25764610 messages 25160 MB payloads
932721.4 msgs/sec, 910.8598 MB/sec, totals 26697330 messages 26071 MB payloads
926913.8 msgs/sec, 905.1903 MB/sec, totals 27624245 messages 26976 MB payloads
923224.4 msgs/sec, 901.5853 MB/sec, totals 28547459 messages 27878 MB payloads
901869.9 msgs/sec, 880.7323 MB/sec, totals 29449332 messages 28759 MB payloads
```

`rsocket-aeron-client-tps.sh`
```
877726.9 msgs/sec, 857.1552 MB/sec, totals 6630733 messages 6475 MB payloads
920794.1 msgs/sec, 899.2130 MB/sec, totals 7551526 messages 7374 MB payloads
949155.2 msgs/sec, 926.9093 MB/sec, totals 8500680 messages 8301 MB payloads
931588.0 msgs/sec, 909.7539 MB/sec, totals 9432271 messages 9211 MB payloads
951305.7 msgs/sec, 929.0095 MB/sec, totals 10383576 messages 10140 MB payloads
933889.7 msgs/sec, 912.0017 MB/sec, totals 11317472 messages 11052 MB payloads
943238.7 msgs/sec, 921.1315 MB/sec, totals 12260711 messages 11973 MB payloads
878650.8 msgs/sec, 858.0575 MB/sec, totals 13139363 messages 12831 MB payloads
895582.6 msgs/sec, 874.5923 MB/sec, totals 14034936 messages 13705 MB payloads
941866.4 msgs/sec, 919.7914 MB/sec, totals 14976804 messages 14625 MB payloads
922469.2 msgs/sec, 900.8489 MB/sec, totals 15899271 messages 15526 MB payloads
935771.9 msgs/sec, 913.8397 MB/sec, totals 16837884 messages 16443 MB payloads
```

`rsocket-aeron-client-simple-tps.sh`
```
895233.5 msgs/sec, 874.2514 MB/sec, totals 5366075 messages 5240 MB payloads
931679.8 msgs/sec, 909.8435 MB/sec, totals 6297756 messages 6150 MB payloads
953895.3 msgs/sec, 931.5383 MB/sec, totals 7251648 messages 7081 MB payloads
895885.8 msgs/sec, 874.8885 MB/sec, totals 8147537 messages 7956 MB payloads
914330.8 msgs/sec, 892.9011 MB/sec, totals 9061866 messages 8849 MB payloads
944791.3 msgs/sec, 922.6478 MB/sec, totals 10006659 messages 9772 MB payloads
922439.3 msgs/sec, 900.8196 MB/sec, totals 10929095 messages 10672 MB payloads
890938.3 msgs/sec, 870.0569 MB/sec, totals 11820034 messages 11543 MB payloads
911943.7 msgs/sec, 890.5700 MB/sec, totals 12731977 messages 12433 MB payloads
944100.9 msgs/sec, 921.9735 MB/sec, totals 13676072 messages 13355 MB payloads
912907.2 msgs/sec, 891.5110 MB/sec, totals 14588984 messages 14247 MB payloads
911598.1 msgs/sec, 890.2325 MB/sec, totals 15500579 messages 15137 MB payloads
939291.3 msgs/sec, 917.2767 MB/sec, totals 16439862 messages 16054 MB payloads
```